refactor: move ValueMapper to feel module
