+++ feel/src/main/scala/io/camunda/zeebe/feel/impl/Loggers.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Zeebe Community License 1.1. You may not use this file
+ * except in compliance with the Zeebe Community License 1.1.
+ */
+package io.camunda.zeebe.feel.impl;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public final class Loggers {
+  public static final Logger LOGGER = LoggerFactory.getLogger("io.camunda.zeebe.feel");
+}
