+++ feel/src/main/scala/io/camunda/zeebe/feel/impl/MessagePackValueMapper.scala
@@ -5,9 +5,9 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.el.impl.feel
+package io.camunda.zeebe.feel.impl
 
-import io.camunda.zeebe.el.impl.Loggers.LOGGER
+import io.camunda.zeebe.feel.impl.Loggers.LOGGER
 import io.camunda.zeebe.msgpack.spec.{MsgPackReader, MsgPackToken, MsgPackType}
 import io.camunda.zeebe.util.buffer.BufferUtil.bufferAsString
 import org.agrona.DirectBuffer
